Add component documentation.
