+++ catalog/dropdown/documentation.md
@@ -1,0 +1,13 @@
+This documentation is automatically generated from jsdoc comments.
+
+```react
+noSource: true
+---
+<DocgenTable component={Dropdown} />
+```
+
+```react
+noSource: true
+---
+<DocgenTable component={DropdownToggle} />
+```
