+++ catalog/dropdown/item-documentation.md
@@ -1,0 +1,17 @@
+This documentation is automatically generated from jsdoc comments.
+
+```react
+noSource: true
+---
+<DocgenTable component={MenuItem} />
+```
+
+```react
+noSource: true
+---
+<DocgenTable component={MenuCheckbox} />
+```
+
+### MenuSeparator
+
+"Items in a menu may be divided into groups by placing an element with a role of separator between groups. For example, this technique should be used when a menu contains a set of menuitemradio items." [wai-aria-practices](www.w3.org/TR/wai-aria-practices-1.1/#menubutton)
