+++ components/dropdown/dropdown-children/index.js
@@ -1,0 +1,4 @@
+export { MenuItem } from './menu-item';
+export { MenuCheckbox } from './menu-checkbox';
+
+export { MenuSeparator } from '../styled';
