+++ components/dropdown/index.js
@@ -1,3 +1,4 @@
 export { Dropdown } from './component';
-export { DropdownToggle, MenuItem, MenuSeparator, MenuCheckbox } from './dropdown-children';
+export { MenuItem, MenuSeparator, MenuCheckbox } from './dropdown-children';
 export { DropdownMenu } from './dropdown-menu';
+export { DropdownToggle } from './dropdown-toggle';
